# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rishika Nagpal & Associates - Law Firm Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  Admin: "Admin",
  Advocate: "Advocate",
  Intern: "Intern",
} as const;

export interface User {
  id: number;
  name: string;
  barId?: string | null;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  Admin: "Admin",
  Advocate: "Advocate",
  Intern: "Intern",
} as const;

export interface CreateUserRequest {
  name: string;
  barId?: string | null;
  email: string;
  role: CreateUserRequestRole;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  Admin: "Admin",
  Advocate: "Advocate",
  Intern: "Intern",
} as const;

export interface UpdateUserRequest {
  name?: string;
  barId?: string | null;
  email?: string;
  role?: UpdateUserRequestRole;
}

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  Active: "Active",
  Pending: "Pending",
  Closed: "Closed",
  Adjourned: "Adjourned",
} as const;

export interface Case {
  id: number;
  caseTitle: string;
  caseNumber: string;
  courtName: string;
  nextDate?: string | null;
  status: CaseStatus;
  createdAt: string;
}

export type CreateCaseRequestStatus =
  (typeof CreateCaseRequestStatus)[keyof typeof CreateCaseRequestStatus];

export const CreateCaseRequestStatus = {
  Active: "Active",
  Pending: "Pending",
  Closed: "Closed",
  Adjourned: "Adjourned",
} as const;

export interface CreateCaseRequest {
  caseTitle: string;
  caseNumber: string;
  courtName: string;
  nextDate?: string | null;
  status: CreateCaseRequestStatus;
}

export type UpdateCaseRequestStatus =
  (typeof UpdateCaseRequestStatus)[keyof typeof UpdateCaseRequestStatus];

export const UpdateCaseRequestStatus = {
  Active: "Active",
  Pending: "Pending",
  Closed: "Closed",
  Adjourned: "Adjourned",
} as const;

export interface UpdateCaseRequest {
  caseTitle?: string;
  caseNumber?: string;
  courtName?: string;
  nextDate?: string | null;
  status?: UpdateCaseRequestStatus;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  Pending: "Pending",
  In_Progress: "In Progress",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface Task {
  id: number;
  taskDescription: string;
  fee?: string | null;
  location?: string | null;
  status: TaskStatus;
  createdAt: string;
}

export type CreateTaskRequestStatus =
  (typeof CreateTaskRequestStatus)[keyof typeof CreateTaskRequestStatus];

export const CreateTaskRequestStatus = {
  Pending: "Pending",
  In_Progress: "In Progress",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface CreateTaskRequest {
  taskDescription: string;
  fee?: string | null;
  location?: string | null;
  status: CreateTaskRequestStatus;
}

export type UpdateTaskRequestStatus =
  (typeof UpdateTaskRequestStatus)[keyof typeof UpdateTaskRequestStatus];

export const UpdateTaskRequestStatus = {
  Pending: "Pending",
  In_Progress: "In Progress",
  Completed: "Completed",
  Cancelled: "Cancelled",
} as const;

export interface UpdateTaskRequest {
  taskDescription?: string;
  fee?: string | null;
  location?: string | null;
  status?: UpdateTaskRequestStatus;
}