# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rishika Nagpal & Associates - Law Firm Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management
  - name: cases
    description: Case management
  - name: tasks
    description: Task management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted

  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Case not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseRequest"
      responses:
        "200":
          description: Case updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Case deleted

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        barId:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [Admin, Advocate, Intern]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt

    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        barId:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [Admin, Advocate, Intern]
      required:
        - name
        - email
        - role

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        barId:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [Admin, Advocate, Intern]

    Case:
      type: object
      properties:
        id:
          type: integer
        caseTitle:
          type: string
        caseNumber:
          type: string
        courtName:
          type: string
        nextDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [Active, Pending, Closed, Adjourned]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseTitle
        - caseNumber
        - courtName
        - status
        - createdAt

    CreateCaseRequest:
      type: object
      properties:
        caseTitle:
          type: string
        caseNumber:
          type: string
        courtName:
          type: string
        nextDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [Active, Pending, Closed, Adjourned]
      required:
        - caseTitle
        - caseNumber
        - courtName
        - status

    UpdateCaseRequest:
      type: object
      properties:
        caseTitle:
          type: string
        caseNumber:
          type: string
        courtName:
          type: string
        nextDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [Active, Pending, Closed, Adjourned]

    Task:
      type: object
      properties:
        id:
          type: integer
        taskDescription:
          type: string
        fee:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        status:
          type: string
          enum: [Pending, In Progress, Completed, Cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taskDescription
        - status
        - createdAt

    CreateTaskRequest:
      type: object
      properties:
        taskDescription:
          type: string
        fee:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        status:
          type: string
          enum: [Pending, In Progress, Completed, Cancelled]
      required:
        - taskDescription
        - status

    UpdateTaskRequest:
      type: object
      properties:
        taskDescription:
          type: string
        fee:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        status:
          type: string
          enum: [Pending, In Progress, Completed, Cancelled]