# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rishika Nagpal & Associates - Law Firm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from "./caseStatus";

export interface Case {
  id: number;
  caseTitle: string;
  caseNumber: string;
  courtName: string;
  nextDate?: string | null;
  status: CaseStatus;
  createdAt: Date;
}