# lib/api-zod/src/generated/types/createCaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rishika Nagpal & Associates - Law Firm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCaseRequestStatus } from "./createCaseRequestStatus";

export interface CreateCaseRequest {
  caseTitle: string;
  caseNumber: string;
  courtName: string;
  nextDate?: string | null;
  status: CreateCaseRequestStatus;
}