# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rishika Nagpal & Associates - Law Firm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  taskDescription: string;
  fee?: string | null;
  location?: string | null;
  status: TaskStatus;
  createdAt: Date;
}