# lib/api-zod/src/generated/types/updateCaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rishika Nagpal & Associates - Law Firm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCaseRequestStatus } from "./updateCaseRequestStatus";

export interface UpdateCaseRequest {
  caseTitle?: string;
  caseNumber?: string;
  courtName?: string;
  nextDate?: string | null;
  status?: UpdateCaseRequestStatus;
}